+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Unicode chars breaks the view [#71]
+
 ### Added
 
 - Telescope support [#61]
@@ -118,3 +122,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#40]: https://github.com/soywod/himalaya/issues/40
 [#41]: https://github.com/soywod/himalaya/issues/41
 [#61]: https://github.com/soywod/himalaya/issues/61
+[#71]: https://github.com/soywod/himalaya/issues/71